refactor!: extract Tab keydown handler into opt-in module
Tab interception, root dummy inputs, and the \`DummyInputManager\` phantom
helpers used to be eagerly installed by every \`createTabster(win)\` call,
even for consumers that only used \`findFocusable\` or feature APIs. The
machinery is now in \`src/Tab.ts\` and only enters the bundle when the
consumer calls \`getRootDummyInputs(tabster)\`.

Specifically:

- New \`src/Tab.ts\` owns \`onKeyDown\`, \`findNextTabbable\`,
  \`getUncontrolledCompletelyContainer\`, the \`isTabbing\` flag, and
  \`installTabKeyHandler(tabster, getWindow)\`. \`FocusedElementState.findNextTabbable\`
  / \`FocusedElementState.isTabbing\` are gone — Mover/Groupper/Modalizer
  dummy managers import \`findNextTabbable\` directly; Mover reads
  \`isTabbing()\` as a function.
- \`Tabster.ts\` no longer imports \`getRootDummyInputs\` and no longer
  auto-calls it. \`controlTab\` defaults to \`false\` (slim baseline).
  \`FocusedElementState.forgetMemorized\` becomes \`_forgetMemorized\` (drops
  the const wrapper).
- \`getRootDummyInputs(tabster)\` now also installs the keydown handler;
  root dummies + keyhandler installation are gated on
  \`controlTab || rootDummyInputs\` so callers can use the function purely
  to register per-feature dummy infrastructure (factories + observer)
  needed by Mover/Groupper/Modalizer dummies in uncontrolled mode.
- Test wrapper (\`tests/test-setup.js\`) defaults \`controlTab: true\` and
  auto-calls \`getRootDummyInputs\` so existing tests that create their
  own Tabster instance still get controlled Tab behaviour.

Bundle savings vs. baseline (monosize):

  createTabster (core)            -12.68 kB  (-41%)
  findAllFocusable                 -9.63 kB  (-31%)
  getRestorer                     -12.34 kB  (-37%)
  getObservedElement              -12.74 kB  (-36%)
  getOutline                      -12.68 kB  (-34%)
  getDeloser                       -8.12 kB  (-21%)
  getMover                         -4.74 kB  (-11%)
  getGroupper                      -4.58 kB  (-12%)
  getModalizer                     -4.03 kB  (-10%)
  getCrossOrigin                   -3.79 kB  (-4.2%)
  all exports                      -3.83 kB  (-4.1%)
  getModalizer (with dummy inputs) -0.56 kB  (-1.4%)

\`getModalizerWithDummyInputs\` is the "consumer asks for Tab" baseline:
the fixture explicitly imports \`getRootDummyInputs\`, so most of the
machinery is back in. The 0.56 kB Δ there comes from the FocusedElement
api-method moves in the prior commit.

Apps that want the previous "Tab just works" behaviour should pair
\`createTabster(win)\` with \`getRootDummyInputs(tabster)\`.
