+++ src/get/getRootDummyInputs.ts
@@ -5,22 +5,25 @@
 
 import { DummyInputManager, createDummyInputObserver } from "../DummyInput.js";
 import { createRootDummyManager } from "../RootDummyManager.js";
+import { installTabKeyHandler } from "../Tab.js";
 import type * as Types from "../Types.js";
 
 /**
- * Opt the Tabster instance into root-level dummy-input behaviour.
- * Registers the root dummy-manager factory plus the dummy-input observer
- * and the `moveOutOfRoot` routing, then triggers root dummy creation when
- * `controlTab` or `rootDummyInputs` is set. Per-part dummy factories
- * (Mover, Groupper, Modalizer) are registered by their respective `getX`
- * factories, so consumers only pay for the dummy code of features they
- * actually use.
+ * Opt the Tabster instance into Tab-key control + root-level dummy
+ * inputs.
  *
- * `createTabster` calls this automatically when `controlTab` (default
- * `true`) or `rootDummyInputs` is on, so the default
- * `createTabster(window)` "just works." Apps that opt out
- * (`createTabster(win, { controlTab: false })` without calling this) keep
- * the slim, dummy-free baseline.
+ * Registers the root dummy-manager factory, the dummy-input observer,
+ * the `moveOutOfRoot` routing, and the Tab-key keydown handler — all of
+ * which used to be eagerly installed by `createTabster`. They now only
+ * enter the bundle when the consumer calls this function.
+ *
+ * Per-part dummy factories (Mover, Groupper, Modalizer) are registered
+ * by their respective `getX` factories, so consumers only pay for the
+ * dummy code of features they actually use.
+ *
+ * Default `createTabster(win)` is the slim baseline (no Tab control, no
+ * root dummies, no phantom-dummy machinery). Pair it with
+ * `getRootDummyInputs(tabster)` for "Tab just works" behaviour.
  */
 export function getRootDummyInputs(tabster: Types.Tabster): void {
     const tabsterCore = tabster.core;
@@ -61,11 +64,25 @@ export function getRootDummyInputs(tabster: Types.Tabster): void {
             }
         };
 
-        // Apply to existing roots when the controlTab/rootDummyInputs flags
-        // say root should host dummies; per-part Mover/Groupper/Modalizer
-        // dummies are picked up lazily via the registered factories when
-        // each part instance is constructed.
+        // The Tab-key handler and root dummy inputs are gated on the
+        // `controlTab` / `rootDummyInputs` flags so callers can use this
+        // function purely to register the per-feature dummy infrastructure
+        // (factories + observer) needed by Mover/Groupper/Modalizer dummies
+        // in uncontrolled mode without paying for keyhandler/root-dummy
+        // wiring they don't want.
         if (tabsterCore.controlTab || tabsterCore.rootDummyInputs) {
+            // Install the Tab-key handler. Lives in src/Tab.ts so its
+            // dependencies (`findNextTabbable`, `DummyInputManager`'s phantom
+            // helpers, the keydown wiring) only enter the bundle here.
+            const stopTabKeyHandler = installTabKeyHandler(
+                tabsterCore,
+                tabsterCore.getWindow
+            );
+            tabsterCore.disposers.add({ dispose: stopTabKeyHandler });
+
+            // Apply to existing roots; per-part Mover/Groupper/Modalizer
+            // dummies are picked up lazily via the registered factories when
+            // each part instance is constructed.
             tabsterCore.root.addDummyInputs();
         }
     }
